+++ src/_P148_POWRxxD_THR3xxD.ino
@@ -62,7 +62,7 @@ boolean Plugin_148(uint8_t function, struct EventStruct *event, String& string)
       };
       constexpr size_t nrElements = sizeof(values) / sizeof(values[0]);
 
-      for (int i = 0; i < nrElements; ++i) {
+      for (size_t i = 0; i < nrElements; ++i) {
         if (i != 0) { addHtml(event->String1); }
         addHtml(labels[i]);
         addHtml(F(":&nbsp;"));
